pyproject.toml
[project]
name = "claude-tracker"
version = "0.1.0"
description = "Windows desktop widget showing Claude Code usage limits in real-time"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.0",

    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
]

[project.scripts]
claude-tracker = "claude_tracker.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_tracker"]
